Factor out service plan summary
